+++ Code/CanSat/Transmitting.h
@@ -38,7 +38,7 @@
 #include "RFM69.h"
 
 void transmitting_init(bool use_radio, bool use_sms, bool use_email);
-void transmitting_send(int data[], int size);
+void transmitting_send(long int data[], int size);
 void transmitting_send(char *message, int len);
 
 #endif //CANSAT_TRANSMITTING_H
